refactor(StringCalculator): NumberList 생성로직 변경 피드백 반영
