+++ src/main/java/stringcalculator/NumberList.java
@@ -1,8 +1,6 @@
 package stringcalculator;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class NumberList {
     private final List<Number> numberList;
@@ -11,19 +9,8 @@ public class NumberList {
         this.numberList = numberList;
     }
 
-    public static NumberList of(String[] numbers) {
-        List<Number> numbersList = createNumberList(numbers);
-        return new NumberList(numbersList);
-    }
-
-    private static List<Number> createNumberList(String[] numbers) {
-        return Arrays.stream(numbers)
-                .map(NumberList::convertNumber)
-                .collect(Collectors.toList());
-    }
-
-    private static Number convertNumber(String number) {
-        return Number.of(Integer.parseInt(number));
+    public static NumberList of(List<Number> numberList) {
+        return new NumberList(numberList);
     }
 
     public int sum() {
