+++ src/main/java/stringcalculator/NumberListParser.java
@@ -1,7 +1,10 @@
 package stringcalculator;
 
+import java.util.Arrays;
+import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 public class NumberListParser {
     private static final Pattern PATTERN = Pattern.compile("//(.)\n(.*)");
@@ -14,8 +17,19 @@ public class NumberListParser {
         if (matcher.find()) {
             String customDelimiter = matcher.group(CUSTOM_DELIMITER_NO);
             String[] numberStringArray = matcher.group(CUSTOM_DELIMITER_NUMBERS_NO).split(customDelimiter);
-            return NumberList.of(numberStringArray);
+            List<Number> numberList = createNumberList(numberStringArray);
+            return NumberList.of(numberList);
         }
-        return NumberList.of(input.split(COMMA_OR_COLON));
+        return NumberList.of(createNumberList(input.split(COMMA_OR_COLON)));
+    }
+
+    private static List<Number> createNumberList(String[] numbers) {
+        return Arrays.stream(numbers)
+                .map(NumberListParser::convertNumber)
+                .collect(Collectors.toList());
+    }
+
+    private static Number convertNumber(String number) {
+        return Number.of(Integer.parseInt(number));
     }
 }
